add transaction history item
