+++ src/locales/en/common.json
@@ -114,7 +114,7 @@
     "VOLUME_IN_24H": "Volume in 24 H",
     "TOTAL_AMOUNT": "Total Amount",
     "HOLDERS": "Holders",
-    "TTOTAL_TRANSFERS": "Total Transfers",
+    "TOTAL_TRANSFERS": "Total Transfers",
     "RED_FLAG": "Red Flag"
   },
   "BLOCK_DETAIL_PAGE": {
